requirements.txt
streamlit==1.30.0
plotly==5.9.0
pandas==2.0.3
numpy==1.24.3
folium==0.15.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
haversine==2.8.1
streamlit-folium==0.18.0
Pillow==9.4.0